renamed moved updated files - files are renamed and moved - imports are updated - new componenets are added
